+++ packages/av-canvas/CHANGELOG.md
@@ -7,13 +7,66 @@
 - Updated dependencies
   - @webav/av-cliper@0.0.0-playback-rate-20240823090827
 
+## 0.15.5
+
 ## 0.14.15
 
 ### Patch Changes
 
-- Updated dependencies [7980ebd]
-- Updated dependencies [a9d810a]
-  - @webav/av-cliper@0.14.15
+- Updated dependencies [e3b9a74]
+  - @webav/av-cliper@0.15.5
+
+## 0.15.4
+
+### Patch Changes
+
+- b50da15: feat: support captureImage for AVCanvas
+  - @webav/av-cliper@0.15.4
+
+## 0.15.3
+
+### Patch Changes
+
+- Updated dependencies [e9b5351]
+  - @webav/av-cliper@0.15.3
+
+## 0.15.2
+
+### Patch Changes
+
+- Updated dependencies [21ffb44]
+  - @webav/av-cliper@0.15.2
+
+## 0.15.1
+
+### Patch Changes
+
+- Updated dependencies [bd37284]
+  - @webav/av-cliper@0.15.1
+
+## 0.15.0
+
+### Patch Changes
+
+- 8d76376: chore: merge bugfix
+- Updated dependencies [8d76376]
+- Updated dependencies [d935bdc]
+  - @webav/av-cliper@0.15.0
+
+## 0.15.0-beta.1
+
+### Patch Changes
+
+- chore: merge bugfix
+- Updated dependencies
+  - @webav/av-cliper@0.15.0-beta.1
+
+## 0.15.0-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @webav/av-cliper@0.15.0-beta.0
 
 ## 0.14.14