+++ packages/av-recorder/CHANGELOG.md
@@ -2,10 +2,66 @@
 
 ## 0.0.0-playback-rate-20240823090827
 
+## 0.15.5
+
+### Patch Changes
+
+- Updated dependencies [e3b9a74]
+  - @webav/av-cliper@0.15.5
+
+## 0.15.4
+
+### Patch Changes
+
+- @webav/av-cliper@0.15.4
+
+## 0.15.3
+
+### Patch Changes
+
+- Updated dependencies [e9b5351]
+  - @webav/av-cliper@0.15.3
+
+## 0.15.2
+
+### Patch Changes
+
+- Updated dependencies [21ffb44]
+  - @webav/av-cliper@0.15.2
+
+## 0.15.1
+
+### Patch Changes
+
+- Updated dependencies [bd37284]
+  - @webav/av-cliper@0.15.1
+
+## 0.15.0
+
+### Patch Changes
+
+- 8d76376: chore: merge bugfix
+- Updated dependencies [8d76376]
+- Updated dependencies [d935bdc]
+  - @webav/av-cliper@0.15.0
+
+## 0.15.0-beta.1
+
+### Patch Changes
+
+- chore: merge bugfix
+- Updated dependencies
+  - @webav/av-cliper@0.15.0-beta.1
+
+## 0.15.0-beta.0
+
 ### Patch Changes
 
 - Updated dependencies
-  - @webav/av-cliper@0.0.0-playback-rate-20240823090827
+  <<<<<<< HEAD
+  - # @webav/av-cliper@0.0.0-playback-rate-20240823090827
+  - @webav/av-cliper@0.15.0-beta.0
+    > > > > > > > origin/main
 
 ## 0.14.15